# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Amadeus Node HTTP API
  description: |
    REST API for the Amadeus blockchain node, providing access to blockchain data,
    peer information, wallet operations, and smart contract interactions.

    This API enables developers to:
    - Query blockchain state and transactions
    - Manage wallet balances and transfers
    - Interact with smart contracts
    - Monitor network peers and validators

    ## Authentication
    Most endpoints are publicly accessible for reading blockchain data.
    Transaction submission requires proper cryptographic signatures.

    ## Base URL
    The API is typically available at `http://localhost` (port 80 by default, configurable via HTTP_PORT env var) when running a local node.
  version: 1.1.9
  contact:
    name: Amadeus Project
    url: https://github.com/amadeus-robot/node
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://72.9.144.110
    description: Amadeus seed node (primary)
  - url: http://167.235.169.185
    description: Amadeus seed node (secondary)
  - url: http://37.27.238.30
    description: Amadeus seed node (tertiary)
  - url: http://localhost
    description: Local development server (HTTP_PORT=80)
...
paths:
...
  /api/wallet/balance/{publicKey}:
    get:
      tags: [wallet]
      summary: Get wallet balance for default token
      description: Retrieves the AMA token balance for a specific public key
      operationId: getWalletBalance
      parameters:
        - name: publicKey
          in: path
          required: true
          description: Base58-encoded public key
          schema:
            type: string
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X GET "http://72.9.144.110/api/wallet/balance/{publicKey}"
  /api/wallet/balance/{publicKey}/{symbol}:
    get:
      tags: [wallet]
      summary: Get wallet balance for specific token
      description: Retrieves the balance for a specific token symbol
      operationId: getWalletBalanceBySymbol
      parameters:
        - name: publicKey
          in: path
          required: true
          description: Base58-encoded public key
          schema:
            type: string
        - name: symbol
          in: path
          required: true
          description: Token symbol (e.g., AMA, USDT)
          schema:
            type: string
            example: AMA
      responses:
        '200':
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /api/wallet/balance_all/{publicKey}:
    get:
      tags: [wallet]
      summary: Get all token balances
      description: Retrieves balances for all tokens held by a specific public key
      operationId: getAllWalletBalances
      parameters:
        - name: publicKey
          in: path
          required: true
          description: Base58-encoded public key
          schema:
            type: string
      responses:
        '200':
          description: All balances retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllBalancesResponse'
  /api/tx/submit:
    post:
      tags: [transaction]
      summary: Submit transaction (POST body)
      description: |
        Submits a signed transaction to the network for processing.
        The transaction must be properly signed and formatted.
      operationId: submitTransaction
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: Packed and signed transaction data
          text/plain:
            schema:
              type: string
              description: Base58-encoded packed transaction
      responses:
        '200':
          description: Transaction submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionSubmitResponse'
        '400':
          description: Invalid transaction format or signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codeSamples:
        - lang: curl
          label: cURL (Binary)
          source: |
            curl -X POST \
              -H "Content-Type: application/octet-stream" \
              --data-binary @transaction.bin \
              "http://72.9.144.110/api/tx/submit"
        - lang: curl
          label: cURL (Base58)
          source: |
            curl -X POST \
              -H "Content-Type: text/plain" \
              -d "7EVUJfpnEqK32KrzUAaR4Yf26KgT66AWwW63xSHw5mbgdQhV8iwL1NkHMAqTi5Hv3h" \
              "http://72.9.144.110/api/tx/submit"
  /api/tx/submit/{txPackedBase58}:
    get:
      tags: [transaction]
      summary: Submit transaction (GET URL)
      description: Submits a Base58-encoded transaction via URL parameter
      operationId: submitTransactionViaUrl
      parameters:
        - name: txPackedBase58
          in: path
          required: true
          description: Base58-encoded packed transaction
          schema:
            type: string
      responses:
        '200':
          description: Transaction submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionSubmitResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum: [not_found, invalid_format, network_error]
          description: Error type
      required: [error]
...
    BalanceResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok]
        balance:
          $ref: '#/components/schemas/Balance'
      required: [error, balance]
    AllBalancesResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok]
        balances:
          type: array
          items:
            $ref: '#/components/schemas/Balance'
      required: [error, balances]
    Balance:
      type: object
      properties:
        symbol:
          type: string
          description: Token symbol
          example: "AMA"
        flat:
          type: integer
          description: Balance in smallest unit (atoms)
          example: 1000000000000
        float:
          type: number
          description: Human-readable balance
          example: 1000.0
      required: [symbol, flat, float]
    TransactionSubmitResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok, invalid_signature, insufficient_funds, network_error]
        tx_hash:
          type: string
          description: Transaction hash if successful
        message:
          type: string
          description: Additional information
...