# http/src/routes/api/transaction.rs
use crate::models::*;
use ama_core::Context;
use axum::{
    body::Bytes,
    extract::{Path, State},
    http::StatusCode,
    Json,
};
use std::sync::Arc;
use utoipa;

// Router function not needed - routes defined in mod.rs

#[utoipa::path(
    post,
    path = "/api/tx/submit",
    summary = "Submit transaction",
    description = "Submits a transaction to the network for processing",
    request_body = Vec<u8>,
    responses(
        (status = 200, description = "Transaction submitted successfully", body = TransactionSubmitResponse),
        (status = 400, description = "Invalid transaction data", body = TransactionSubmitResponse)
    ),
    tag = "transaction"
)]
pub async fn submit_transaction(
    State(_ctx): State<Arc<Context>>,
    body: Bytes,
) -> Result<Json<TransactionSubmitResponse>, StatusCode> {
    // placeholder implementation - would process and submit transaction to mempool

    if body.is_empty() {
        return Ok(Json(TransactionSubmitResponse::error("invalid_signature")));
    }

    // simulate transaction processing
    let tx_hash = format!("tx_{:x}", body.len());

    Ok(Json(TransactionSubmitResponse::ok(tx_hash)))
}

#[utoipa::path(
    get,
    path = "/api/tx/submit/{tx_packed_base58}",
    summary = "Submit transaction via URL",
    description = "Submits a Base58-encoded transaction via URL parameter",
    responses(
        (status = 200, description = "Transaction submitted successfully", body = TransactionSubmitResponse)
    ),
    params(
        ("tx_packed_base58" = String, Path, description = "Base58-encoded packed transaction", example = "2xVqgdkuLqjGkBqwhbiVqArhsrRrgF4RjmD3AuNpneiXGcMY4aE8gj6zBPYF6KhbJUBvggsTnfhpAVrTgqnCtJWe")
    ),
    tag = "transaction"
)]
pub async fn submit_transaction_via_url(
    State(_ctx): State<Arc<Context>>,
    Path(tx_packed_base58): Path<String>,
) -> Json<TransactionSubmitResponse> {
    // placeholder implementation - would decode base58 and submit transaction

    if tx_packed_base58.is_empty() {
        return Json(TransactionSubmitResponse::error("invalid_format"));
    }

    // simulate base58 decoding and transaction processing
    match bs58::decode(&tx_packed_base58).into_vec() {
        Ok(tx_data) => {
            let tx_hash = format!("url_tx_{:x}", tx_data.len());
            Json(TransactionSubmitResponse::ok(tx_hash))
        }
        Err(_) => Json(TransactionSubmitResponse::error("invalid_format")),
    }
}